requirements.txt
torch
torchvision
timm
numpy
Pillow
scikit-learn
